requirements.txt
Keras==2.4.3
matplotlib==3.4.1
numpy==1.18.5
pandas==1.2.2
streamlit==0.84.0
scikit_image==0.18.2
Pillow==8.3.1
skimage==0.0
